% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farmacia}
\alias{farmacia}
\title{Biotec
I dati: Un dataset che contiene informazioni su 41 studenti di Biotec}
\format{A data frame with 41 rows and 8 variabili:
\describe{
  \item{SESSO}{M/F}
  \item{PESO}{peso in kg}
  \item{ALTEZZA}{Altezza in metri}
  \item{OCCHI}{Colore degli occhi}
  \item{CAPELLI}{Colore dei capelli}
  \item{SCARPE}{numero di scarpe}
  \item{hMADRE}{Altezza della madre}
  \item{hPADRE}{Altezza del padre}
  
}


Farmacia
I dati: Un dataset che contiene informazioni su 55 studenti di Farmacia}
\usage{
farmacia
}
\description{
Biotec
I dati: Un dataset che contiene informazioni su 41 studenti di Biotec
}
\keyword{datasets}

