% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidinterest}
\alias{kidinterest}
\title{Interesse dei bambini
The Data: Un dataset che contiene informazioni su 478 bambini}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Gender}{Sesso}
  \item{Age}{eta' in anni}
  \item{Race}{Bianco o altro}
  \item{Urban/Rural}{distretto scolastico di tipo rurale, urbano o suburbano}
  \item{School}{eta' in anni}
  \item{Goals}{Gli obiettivi degli studenti erano  1 =Ottenere buone valutazioni, 2 = Essere popolari,
   3 = Essere bravo negli sport}
  \item{Grades}{(1=most important for popularity, 4=least important)}
  \item{Sport}{eta' in anni}
  \item{Looks}{(1=most important for popularity, 4=least important)i}
  \item{Money}{eta' in anni}
}}
\source{
\url{http://lib.stat.cmu.edu/DASL/Datafiles/PopularKids.html}
}
\usage{
kidinterest
}
\description{
Interesse dei bambini
The Data: Un dataset che contiene informazioni su 478 bambini
}
\keyword{datasets}

