% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo}
\alias{meteo}
\title{Dati meteo  a Milano}
\format{Un data frame di 173 righe e 3 variabili:
\describe{
  \item{day }{tempo in giorni}
  \item{living}{moscerini vivi}
  \item{mort.rate}{tasso di mortalita' per ogni giorno}
  ...
}}
\source{
\url{http://lib.stat.cmu.edu/DASL/Datafiles/Medflies.html}
}
\usage{
meteo
}
\description{
Dati meteo  a Milano
}
\keyword{datasets}

